Remove dependency on test target for building operator container.
We already run the tests as a separate job in the CI workflow.
Without this change we run tests twice in our CI workflow.

I've decided to add a run of our tests as part of releasing a new
operator container image. We could also leave it out, but without
this we could end in situations where not all the relevant changes
have been tested together, mainly in scenarios where we have branches
for a PR that are not up to date with master but passes tests.