Allow disabling TLS for ingress.
This is useful if you rely on e.g. AWS NLB's for doing TLS termination,
while the NLB is communicating to ingress-nginx over unencrypted communications.

By default, we will stick to the assumption that TLS should be enabled
on ingress objects. This means TLS will only be disabled for ingress if
explicitly defined as such.

Fixes https://github.com/humio/humio-operator/issues/216