Set annotation on a per-test-process to prevent conflicts when running in parallel.
Without this, if you run the tests in parallel you often end up with
problems where one test finishes early and then the teardown of that
will remove other tests running in parallel.

This is not a problem for our e2e tests yet, as they right now still run
sequentially, but this is part of what is required to make it possible
to run the e2e tests in parallel.