Introduce ConfigError states to all resource types.
We use this state to indicate errors with the user-defined
specifications, such as pointing to a non-existent Humio cluster.

In the future, we might also want to use this state when the user
provides an API token with too few permissions, but for now we
keep it the way it is.

Previously we always called `Status().Update()` even if a field already
had the desired value. This has been changed so we return early and skip
triggering the status update if the value already contains the desired
content.

I've also refactored the HumioView controller to be more similar to the
other controllers as that makes things slightly easier to work with when
trying to make changes across our resource types. We might want to
refactor our Reconcile methods at some point though to make them easier
to read and understand.