Bump to operator-sdk 1.3.0, kustomize 3.8.7 and helm 3.4.2.
Remove unused tools from operator-sdk action.

We've migrated our GitHub Action workflows to run helm/kubectl/etc.
directly from the shell in our workflows instead of running it inside
the operator-sdk action container due to connectivity complications.
Because of this, we already migrated away from using the tools in this
way and is safe to remove. Right now we don't use the operator-sdk
action at all, but we do have a few disabled workflow steps that uses
it, which is why I'm not deleting the action entirely.