Require explicitly defined storage configuration.
Up until the default has been to EmptyDir. It is safer to require the
user to specify which type of storage they want to use. This should
help in scenarios where the user forgot to configure what type of
storage they want to use and thus unknowingly ended up using the
previous default.