Stop using Telepresence to run E2E tests
Instead we run a pod inside the `kind` cluster with the source code and
tools to compile and run the tests inside the cluster itself.

This unfortunately means the tests using `crc` for testing will not
work. We use `kind load docker-image` to load the container where the
tests will be performed within and `crc` doesn't have a similar command.
This means that the `crc` scripts will likely need to either push the
containers to a remote registry, or potentially "docker push" the image
directly to a registry running inside the `crc` cluster.