Sort view connections before comparing
Currently if a view has multiple connections Kubernetes might return the
connection slice in different orders every time. As we compare view connections
with a reflect.DeepEqual that results in updates to views over and over again.
When ever a view is updated in Humio any live queries are refreshed which leads
to flashing dashboards.

This change introduces a mitigation to this behaviour by sorting the existing
and new connection slices before comparing them. This way we only update the
view in Humio if the functional behaviour has changed.