Disable logging stdout from reports
Right now we end up logging everything 3 times, which isn't exactly what
we want. The idea was to make it easy to go from test result to related
log lines, but we shouldn't do that by duplicating all logs.

This disables the 2 additional copies so we only have one copy of
stdout. This does however mean we won't have a direct link from the test
report being logged out to the relevant log entries, but perhaps we can
find another solution to that.