Refactor Helm chart resources
The main change here is to better generate resource names for the
resources created by installing the Helm chart.

This also removes the unnecessary single quote characters around values
in the YAML files, which makes syntax highlighting slightly better for
some editors.

Ideally we'd also change it so that the resource labels would use the
names that can now be overridden, but I'm leaving that out for now.
This also removes the `OPERATOR_NAME` environment variable from the
Deployment resource as we don't use it at all, but perhaps we should
later consider adding it back again and populating it e.g. `fullname`
such that any resources created by this operator installation would
create and only touch resources that it itself installed. This may make
it easier for us to better support running multiple operators on the
same Kubernetes cluster.

Fixes: https://github.com/humio/humio-operator/issues/574