Upgrade Ginkgo, Gomega and Go
This is required for https://github.com/onsi/gomega/releases/tag/v1.23.0
to work.

Our CI job started failing as it uses a newer gomega version compared to
what we have defined in our go.mod file.

I suppose we could also adjust our github workflows so they always stick
to the versions in the go.mod file, but let's keep that as a separate PR
if that is what we want. For now I just bump the issue to fix the
current issue causing CI to fail and marking my PR red.