Stop printing node details for all cluster nodes
Due to the size of the details for large clusters, this ends up being
split into multiple log lines causing log shipping to ship the log line
as multiple events, which breaks parsing of each event since they are
not valid JSON events.

So far, we've not had a case where this log line was important to us, so
let's remove it.