Ensure call to setState doesn't impact what error is returned.
This commit also fixes a bug that could cause the operator to not
retry reconciling an object if e.g. HumioExternalCluster or the
API token secret is gone to construct Humio client config when calling
NewCluster(). In those cases, even if the HumioExternalCluster or
the API token secret is back, the reconcile is never retried.