Treat Delete on entities as successful if entities cannot be found.
This will unblock cases where e.g. finalizers are blocking object
deletion just because the operator keeps trying to delete something that
is returned as EntityNotFound.

There's technically more CRD's we could apply this to, but we'd first
need to extend this EntityNotFound part to those resource types in the
API package.