Fix requeue and finalizers (#838)
* Ensure call to setState doesn't impact what error is returned.

This commit also fixes a bug that could cause the operator to not
retry reconciling an object if e.g. HumioExternalCluster or the
API token secret is gone to construct Humio client config when calling
NewCluster(). In those cases, even if the HumioExternalCluster or
the API token secret is back, the reconcile is never retried.

* Fix example for HumioScheduledSearch

* Refer directly to the humio API package for the query ownership type

* Treat Delete on entities as successful if entities cannot be found.

This will unblock cases where e.g. finalizers are blocking object
deletion just because the operator keeps trying to delete something that
is returned as EntityNotFound.

There's technically more CRD's we could apply this to, but we'd first
need to extend this EntityNotFound part to those resource types in the
API package.