Requeue after incrementHumioClusterPodRevision and refactor podReadyCountByRevision
Requeue after incrementHumioClusterPodRevision fixes a bug where the
operator could in some cases create a new pod using the old revision
as it did not fetch the updated HumioCluster resource before creating
new pods.

Refactor podReadyCountByRevision to extract the logic for what pods to
should be marked as running and ensures we update only the pods that match
the expected pod revision as running, as any older pod revisions should
already be handled during previous iterations.