Add GitHub Action workflow to run golangci-lint.
This also fixes most issues flagged by the default list of linters that
are enabled.

For now, I decided to ignore the handful of places where the gocyclo
rule flags issues, so that we can keep the linter enabled to prevent
us from introducing more code that would be called out by gocyclo. If we
end up fixing those handful of places, we can always remove the nolint
marker.