Add requeue-period flag.
This flag allows operators to manage the default requeue period of the
Humio* CRs managed. In deployments where there are sufficient role
restrictions it should be possible to prevent competing changes to
Logscale entities via the web interface or other GraphQL clients. Thus
the requeue period does not need to be such a short period.