Fix UpdateGroup()
The lookup name / external mapping name is an optional through graphql
API's and return null if not set. To unset, it is necessary to
explicitly set it to the empty string as passing null to the graphql
mutation for updating the group will simply skip/ignore updates to the
external mapping name.