Introduce HumioGroup CRD (#976)
* add HumioGroup CRUD to operator

* handle role assignments for HumioGroups

* bump version number

* fix a state usage

* Keep version as-is. We will update it later when we're ready to push a new release

* HumioGroup: Allow configurable RequeueAfter as other CRD's

* Remove view group assignment from HumioGroup

This will be introduced as part of the role CRD's instead, which will
follow later.

* Remove unused reconcile.Request parameter from Group functions

* Keep version 0.28.2 for now. We will release later

* Fix UpdateGroup()

The lookup name / external mapping name is an optional through graphql
API's and return null if not set. To unset, it is necessary to
explicitly set it to the empty string as passing null to the graphql
mutation for updating the group will simply skip/ignore updates to the
external mapping name.

---------

Co-authored-by: Bowen Sun <bowens@stripe.com>